Refactor out hex2rgb into a utils file
